src/requirements.txt
alembic==0.9.5
amqp==2.2.1
aniso8601==1.2.1
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cache==0.13.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.13
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
